Rimozione metodi e attributi non più utilizzati

+++ headers/Item.h
@@ -16,8 +16,8 @@ public:
     virtual void setName(const std::string &name);
     float getBasePrice() const;
     virtual void setBasePrice(float basePrice);
-    float getQuantity() const;
-    virtual void setQuantity(float quantity);
+    int getQuantity() const;
+    virtual void setQuantity(int quantity);
     const std::string &getCategoryName() const;
     bool isItemInList() const;
     void setItemInList(bool isItemInList);
@@ -34,14 +34,10 @@ public:
 
 protected:
     Item(std::string name, float pricePerUnit, float quantity, bool discounted = false);
-
-    //void isItemInCategoryList(std::string &name, const std::list<std::string> &itemCategories);
     int itemId;
     std::string name;
-    //std::string categoryName;
     float pricePerUnit;
-    float quantity;
-    //std::list<std::string> itemCategories;
+    int quantity;
     bool discounted;
     bool itemInList;
     static int itemCount;
